requirements.txt
esm==3.2.0
esmc_function_classifier==0.0.7
torchao==0.12.0
obonet==1.1.1
pydantic==2.11.0
fastapi==0.115.12
uvicorn==0.34.2